bffApi/pyproject.toml
[tool.poetry]
name = "bffapi"
version = "0.1.0"
description = ""
authors = ["Devlin Duldulao <webmasterdevlin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-rest-framework = "^0.1.0"
circuitbreaker = "^1.3.2"
pyfailsafe = "^0.6.0"
retrying = "^1.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
